fix(infra): satisfy the gcc strict warning gate
Drop a useless size_t cast in shard_index and value-initialise the event
locals in the egress tests so gcc -Wuseless-cast and -Wmaybe-uninitialized
pass under -Werror. Reformat the touched lines with clang-format-18.

+++ include/lob/hash.hpp
@@ -23,7 +23,7 @@ namespace lob {
 // bits as well as the high bits, so masking off the high bits loses no
 // distribution quality.
 [[nodiscard]] constexpr std::size_t shard_index(symbol_id_t sym, std::size_t num_shards) noexcept {
-    return static_cast<std::size_t>(splitmix64(sym) & (num_shards - 1));
+    return splitmix64(sym) & (num_shards - 1);
 }
 
 }  // namespace lob
